Made A_Bossdeath trigger much sooner on plasma deaths for Mancubus/Archnotron
On plasma deaths, these enemies have an abnormally large window before A_BossDeath is called (Almost 5 seconds). If more than one of these enemies are killed in that window,  the special will be activated more than once. Thus, A_BossDeath is now called sooner (once the corpse sits there smoking), giving a window of ~45 tics, similar to vanilla. This bug is present in vanilla also, but the window to trigger it is much smaller (A little over a second).